requirements.txt
PyQt5
clipspy
